+++ lib/plugins/ig/index.js
@@ -1,5 +1,5 @@
 var Plugins = require('../plugin.js'),
-	googleSpreadsheets = require('../../google/googlespreadsheet.js'),
+	googleSpreadsheets = require('../../google/googlespreadsheet-v3-api.js'),
 	GssUtils = require('../../google/gssUtils.js'),
 	_ = require('lodash');
